apps/chatbot/requirements.txt
fastapi==0.115.11
uvicorn==0.33.0
sqlalchemy==2.0.39
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.10.6
python-multipart==0.0.20
pypdf==5.3.0
python-docx==1.1.2
